Add Table in different layout
