+++ frontend/csyllabus/src/app/core/core.component.html
@@ -8,6 +8,7 @@
       <mat-tab label="Comparator">Content 2</mat-tab>
     </mat-tab-group>
   </div>
+  <div><app-explorer-result></app-explorer-result></div>
 </div>
