+++ frontend/csyllabus/src/app/core/explorer/explorer.component.css
@@ -25,52 +25,3 @@
 .form{
     margin: 20px;
 }
-
-/* Table Result Definition */
-/* -*- coding: utf-8 -*-
-# @Author: Adrien Roques
-# @Date:   2017-11-03
-# @Last Modified by:   Adrien Roques
-# @Last Modified time: 2017-11-05 */
-.explorer-table-container {
-  display: flex;
-  flex-direction: column;
-  min-width: 500px;
-  position: relative;
-}
-
-
-.explorer-table-header {
-  display: flex;
-  font-size: 20px;
-  justify-content: center;
-  min-width: 500px;
-}
-
-.mat-form-field {
-  font-size: 14px;
-  flex-grow: 1;
-  margin-left: 32px;
-}
-
-.mat-header-row {
-  background: #00008B;
-}
-
-.mat-table {
-  overflow: auto;
-  text-align: left;  
-  height: 100%;
-}
-
-.mat-header-cell {
-  color: #F8F8FF;
-  font-size: 18px;
-  font-weight: bold;
-}
-
-.mat-cell {
-  display: flex;
-  justify-content: space-between;
-}
-
