+++ frontend/csyllabus/src/app/core/explorer/result/result.component.spec.ts
@@ -1,0 +1,25 @@
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { ResultComponent } from './result.component';
+
+describe('ResultComponent', () => {
+    let component: ResultComponent;
+    let fixture: ComponentFixture<ResultComponent>;
+
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+        declarations: [ResultComponent ]
+    })
+    .compileComponents();
+  }));
+
+  beforeEach(() => {
+      fixture = TestBed.createComponent(ResultComponent);
+      component = fixture.componentInstance;
+      fixture.detectChanges();
+  });
+
+  it('should be created', () => {
+    expect(component).toBeTruthy();
+  });
+});
