+++ frontend/csyllabus/src/app/core/explorer/result/result.result.ts
